Rewrite DB import script to fix bugs with tracks and scrobbles
When importing check to see if the track already exists based on the track name and artist name. Previously tracks with the same name but by different artists were being assigned to the wrong artist.

Christmas also took place and I forgot to commit where I was before. For some reason I decided to swap to pg-promise because I ran into an issue.

I now feel that was unnecessary and was down to my lack of SQL knowledge / unsure of how to progress.

Anyway, now using pg-promise.